# docs/openapi.yaml
openapi: 3.1.0
info:
  title: WebHost API
  version: 0.1.0
  description: >-
    Publish static HTML and get a stable public URL. Uploaded HTML is validated
    server-side and served byte-for-byte from a separate content origin.
  license:
    name: MIT
servers:
  - url: https://api.postplan.tapetide.com
    description: Reference deployment (API)
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: wh_ token
...
paths:
...
  /api/projects/{id}:
    patch:
      summary: Update a project
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }
    delete:
      summary: Archive a project
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      responses: { "200": { description: Archived } }
...